backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
google-genai==0.7.0
google-cloud-aiplatform==1.57.0
websockets>=13.0
python-multipart==0.0.9
pillow==10.3.0
pydantic==2.7.0
